requirements.txt
paho-mqtt>=2.0
redis>=5.0
prometheus-client>=0.20
